added support for setting a global PHPCR ODM translation strategy
